BABL-WORK/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# Caching
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# ML & Analytics
scikit-learn==1.3.2
numpy==1.25.2
torch==2.1.1
reportlab==4.0.4
matplotlib==3.7.2

# File Processing
python-magic==0.4.27
chardet==5.2.0

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Fuzzy Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
